Rename to erc1155-all-balances-of
